pyproject.toml
[tool.poetry]
name = "knowledge-graph"
version = "0.1.0"
description = ""
authors = ["Surajit Ghorai <surajitghorai@promactinfo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.11,<3.12"
neo4j = ">=5.19.0,<5.20.0"
python-dotenv = "^1.0.1"
neomodel = "^5.3.2"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-experimental = "^0.0.65"
langchain-google-genai = "^1.0.10"
unstructured = "^0.15.9"
bs4 = "^0.0.2"
json-repair = "^0.29.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
